Allow using mutable state in the on_tick timeout
Create a small helper class that stores an Rc<RefCell<>> so that the
timeout can access the state.  It seems a little unfortunate that we
have to use RefCell's runtime borrow checks here, even though we know
that the glib main loop is single threaded and there cannot be any other
references when the timeout is running.